fix allow dispatching to streaming from streaming
