+++ crates/polars-stream/src/lib.rs
@@ -1,5 +1,6 @@
 mod async_executor;
 mod async_primitives;
+mod dispatch;
 mod skeleton;
 
 use std::sync::LazyLock;
@@ -7,6 +8,7 @@ use std::sync::LazyLock;
 pub use skeleton::{run_query, visualize_physical_plan};
 
 mod execute;
+pub use dispatch::build_streaming_query_executor;
 pub(crate) mod expression;
 mod graph;
 pub use skeleton::{QueryResult, StreamingQuery};
